content: move repo schema converters to store layer
convertSchemaToIBCustomRepo and convertSchemaToIBPayloadRepo are pure
data transforms between API schemas. They belong in the content subslice,
not in a UI component utility. Update all consumers to import from the
store barrel.
