+++ src/store/slices/wizard/content/index.ts
@@ -2,3 +2,4 @@ export * from './selectors';
 export * from './slice';
 export { initialState as contentState } from './state';
 export * from './types';
+export * from './utilities';
